+++ pclsync/pfstasks.c
@@ -54,7 +54,7 @@ typedef struct {
   char name[];
 } file_history_record;
 
-static psync_tree *folders = PSYNC_TREE_EMPTY;
+psync_tree *folders = PSYNC_TREE_EMPTY;
 static uint64_t psync_local_taskid = UINT64_MAX;
 
 static inline int psync_crypto_is_error(const void *ptr) {
@@ -177,14 +177,11 @@ pfs_task_get_folder_tasks_rdlocked(psync_fsfolderid_t folderid) {
   return NULL;
 }
 
+__attribute__((weak)) void pfstasks_debug_check_folder_consistency(psync_fstask_folder_t *folder) {}
+__attribute__((weak)) void pfs_task_dump_state() {}
+
 void pfs_task_release_folder_tasks_locked(psync_fstask_folder_t *folder) {
-#if IS_DEBUG
-  if ((!!folder->taskscnt) !=
-      (folder->creats || folder->mkdirs || folder->rmdirs || folder->unlinks))
-    pdbg_logf(D_ERROR, "taskcnt=%u, c=%p, m=%p, r=%p, u=%p",
-          (unsigned)folder->taskscnt, folder->creats, folder->mkdirs,
-          folder->rmdirs, folder->unlinks);
-#endif
+  pfstasks_debug_check_folder_consistency(folder);
   if (--folder->refcnt == 0 && !folder->taskscnt) {
     pdbg_logf(D_NOTICE, "releasing folder id %ld", (long int)folder->folderid);
     ptree_del(&folders, &folder->tree);
@@ -2166,48 +2163,3 @@ void pfs_task_init() {
   pfs_upld_init();
 }
 
-#if IS_DEBUG
-
-void pfs_task_dump_state() {
-  psync_fstask_folder_t *folder;
-  psync_fstask_mkdir_t *mk;
-  psync_fstask_rmdir_t *rm;
-  psync_fstask_creat_t *cr;
-  psync_fstask_unlink_t *un;
-  uint32_t cnt;
-  ptree_for_each_element(folder, folders, psync_fstask_folder_t, tree) {
-    pdbg_logf(D_NOTICE, "open folderid %ld taskcnt %u refcnt %u",
-          (long)folder->folderid, (unsigned)folder->taskscnt,
-          (unsigned)folder->refcnt);
-    cnt = 0;
-    ptree_for_each_element(mk, folder->mkdirs, psync_fstask_mkdir_t,
-                                tree) {
-      pdbg_logf(D_NOTICE, "  mkdir %s folderid %ld taskid %lu", mk->name,
-            (long)mk->folderid, (unsigned long)mk->taskid);
-      cnt++;
-    }
-    ptree_for_each_element(rm, folder->rmdirs, psync_fstask_rmdir_t,
-                                tree) {
-      pdbg_logf(D_NOTICE, "  mkdir %s folderid %ld taskid %lu", rm->name,
-            (long)rm->folderid, (unsigned long)rm->taskid);
-      cnt++;
-    }
-    ptree_for_each_element(cr, folder->creats, psync_fstask_creat_t,
-                                tree) {
-      pdbg_logf(D_NOTICE, "  creat %s fileid %ld taskid %lu", cr->name,
-            (long)cr->fileid, (unsigned long)cr->taskid);
-      cnt++;
-    }
-    ptree_for_each_element(un, folder->unlinks, psync_fstask_unlink_t,
-                                tree) {
-      pdbg_logf(D_NOTICE, "  unlink %s fileid %ld taskid %lu", un->name,
-            (long)un->fileid, (unsigned long)un->taskid);
-      cnt++;
-    }
-    if (cnt != folder->taskscnt)
-      pdbg_logf(D_ERROR, "inconsistency found, counted taskcnt %u != taskcnt %u",
-            (unsigned)cnt, (unsigned)folder->taskscnt);
-  }
-}
-
-#endif
