+++ pclsync/pfstasks_internal.h
@@ -1,0 +1,11 @@
+// pfstasks_internal.h - internal state shared between pfstasks.c and debug/pfstasks_debug.c
+// Do NOT include this header from any file other than pfstasks.c and debug/pfstasks_debug.c.
+
+#ifndef PFSTASKS_INTERNAL_H
+#define PFSTASKS_INTERNAL_H
+
+#include "ptree.h"
+
+extern psync_tree *folders;
+
+#endif
