+++ pclsync/psys.c
@@ -13,14 +13,11 @@ static gid_t psync_gid;
 static gid_t *psync_gids;
 static int psync_gids_cnt;
 
+__attribute__((weak)) void psys_debug_abort_on_sqllock(uint64_t millisec) {}
+__attribute__((weak)) void psys_debug_configure_core_dump() {}
+
 static void abort_on_sqllock(uint64_t millisec) {
-#if IS_DEBUG
-  if (psql_locked()) {
-    pdbg_logf(D_CRITICAL, "trying to sleep while holding sql lock, aborting");
-    psql_dump_locks();
-    abort();
-  }
-#endif
+  psys_debug_abort_on_sqllock(millisec);
 }
 
 uid_t psys_get_uid() { return psync_uid; }
@@ -36,15 +33,7 @@ void psys_init() {
   limit.rlim_cur = limit.rlim_max = 2048;
   if (setrlimit(RLIMIT_NOFILE, &limit))
     pdbg_logf(D_ERROR, "setrlimit failed errno=%d", errno);
-#if IS_DEBUG
-  if (getrlimit(RLIMIT_CORE, &limit))
-    pdbg_logf(D_ERROR, "getrlimit failed errno=%d", errno);
-  else {
-    limit.rlim_cur = limit.rlim_max;
-    if (setrlimit(RLIMIT_CORE, &limit))
-      pdbg_logf(D_ERROR, "setrlimit failed errno=%d", errno);
-  }
-#endif
+  psys_debug_configure_core_dump();
   signal(SIGPIPE, SIG_IGN);
   psync_uid = getuid();
   psync_gid = getgid();
